+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [3.8.0] - 2023-03-23
 
 ### Added
-* Add `install-autocomplete` command for installing shell autocompletion
+* Add `install-autocomplete` command for installing shell autocompletion (currently only bash is supported)
 
 ### Fixed
 * Hitting the download endpoint twice in some cases