+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.9.0] - 2023-04-28
+
 ### Added
 * Support for custom file upload timestamp
 
@@ -419,7 +421,8 @@ This version is pinned strictly to `b2-sdk-python==1.16.0` for the same reason.
 * Fix an off-by-one issue when downloading a range of a file (affects library, but not CLI).
 * Better handling of some errors from the B2 service.
 
-[Unreleased]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.8.0...HEAD
+[Unreleased]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.9.0...HEAD
+[3.9.0]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.8.0...v3.9.0
 [3.8.0]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.7.1...v3.8.0
 [3.7.1]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.7.0...v3.7.1
 [3.7.0]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.6.0...v3.7.0